% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sift.nls.R
\name{sift.nls}
\alias{sift.nls}
\title{Perform a nls minimization with a custom function.}
\usage{
sift.nls(fn = NULL, data = NULL, start = NULL, traces = TRUE,
  optimMethod = "Nelder-Mead", chi2max = 6, options = list(yName = "y",
  errName = "err", normal = FALSE))
}
\arguments{
\item{data}{Data should contain a "y" and a "err" columns, the rest are considered as independent variables.}

\item{options}{It is possible to change the name of the dependent variable and the error by modifying this argument.}
}
\description{
Perform a nls minimization with a custom function.
}

